// src/app/@core/models/demo.ts
export class Demo {
  id?: string;
  introduction?: string;
  step1title?: string;
  step1content?: string;
  step2title?: string;
  step2content?: string;
  step3title?: string;
  step3content?: string;
  step4title?: string;
  step4content?: string;
  example?: string;
  why?: string;
  what?: string;
  how?: string;
  whatif?: string;
}